% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_horseshoe.R
\name{grouped_horseshoe}
\alias{grouped_horseshoe}
\title{Grouped horseshoe regression Gibbs sampler.}
\usage{
grouped_horseshoe(X, C, Y, grp_idx, alpha_inits = rep(0, ncol(C)),
  beta_inits = rep(0, ncol(X)), lambda_sq_inits = rep(1, ncol(X)),
  gamma_sq_inits = rep(1, length(unique(grp_idx))), eta_inits = rep(1,
  length(unique(grp_idx))), mu_init = 0, tau_sq_init = 1,
  sigma_sq_init = 1, psi_sq_init = 1, nu_init = 1, n_burn_in = 500,
  n_samples = 1000, n_thin = 1, error_tol = 1e-07, a1 = 0.001,
  b1 = 0.001, a2 = 0.001, b2 = 0.001, verbose = TRUE)
}
\arguments{
\item{X}{A (n x M) matrix of covariates that we want to apply grouped horseshoe shrinkage on.}

\item{C}{A (n x K) matrix of covariates that we want to apply ridge regression shrinkage on (typically adjustment covariates).}

\item{Y}{A (n x 1) column vector of responses.}

\item{grp_idx}{A (1 x M) row vector indicating which group of the J groups the p covariates in X belong to.}

\item{alpha_inits}{A (K x 1) column vector containing initial values for the regression coefficients corresponding to C.}

\item{beta_inits}{A (M x 1) column vector containing initial values for the regression coefficients corresponding to X.}

\item{lambda_sq_inits}{A (M x 1) column vector containing initial values for the local shrinkage parameters.}

\item{gamma_sq_inits}{A (J x 1) column vector containing initial values for the group shrinkage parameters.}

\item{eta_inits}{A (J x 1) column vector containing initial values for the mixing parameters.}

\item{mu_init}{Initial value for the intercept parameter (double).}

\item{tau_sq_init}{Initial value for the global shrinkage parameter (double).}

\item{sigma_sq_init}{Initial value for the residual variance (double).}

\item{psi_sq_init}{Initial value for the variailbity in the alphas (double).}

\item{nu_init}{Initial value for the augmentation variable (double).}

\item{n_burn_in}{The number of burn-in samples (integer).}

\item{n_samples}{The number of posterior draws (integer).}

\item{n_thin}{The thinning interval (integer).}

\item{error_tol}{Parameter that controls numerical stability of the algorithm (double).}

\item{a1}{Shape parameter for the inverse gamma hyperprior on psi_sq (double).}

\item{b1}{Scale parameter for the inverse gamma hyperprior on psi_sq (double).}

\item{a2}{Shape parameter for the gamma hyperprior on the etas (double).}

\item{b2}{Scale parameter for the gamma hyperprior on the etas (double).}

\item{verbose}{Boolean value which indicates whether or not to print the progress of the Gibbs sampler.}
}
\value{
A list containing the posterior draws of (1) the intercept term (mus) (2) the regression coefficients (alphas and betas) (3) the individual shrinkage parameters (lambda_sqs) (4) the group shrinkage parameters (gamma_sqs) (5) the global shrinkage parameter (tau_sqs) (6) the residual error variance (sigma_sqs) (7) the mixing parameter (etas) and (8) the variance of the prior on alpha (psi_sqs). The list also contains the specified hyperparameters (hyperparam_a1, hyperparam_b1, hyperparam_a2, and hyperparam_b2), details regarding the dataset (X, C, Y, grp_idx), and Gibbs sampler details (n_burn_in, n_samples, and n_thin).
}
\description{
Grouped horseshoe regression Gibbs sampler.
}
