% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grouped_horseshoe_gibbs_sampler}
\alias{grouped_horseshoe_gibbs_sampler}
\title{Gibbs sampler for the grouped horseshoe model.}
\usage{
grouped_horseshoe_gibbs_sampler(X, C, Y, grp_idx, grp_size, grp_size_cs,
  alpha_inits, beta_inits, lambda_sq_inits, gamma_sq_inits, eta_inits,
  mu_init = 0, tau_sq_init = 1, sigma_sq_init = 1, psi_sq_init = 1,
  nu_init = 1, n_burn_in = 500L, n_samples = 1000L, n_thin = 1L,
  error_tol = 1e-07, a1 = 0.001, b1 = 0.001, a2 = 0.001,
  b2 = 0.001, verbose = TRUE)
}
\arguments{
\item{X}{A (n x M) matrix of covariates that we want to apply grouped horseshoe shrinkage on.}

\item{C}{A (n x K) matrix of covariates that we want to apply ridge regression shrinkage on (typically adjustment covariates).}

\item{Y}{A (n x 1) column vector of responses.}

\item{grp_idx}{A (1 x M) row vector indicating which group of the J groups the p covariates in X belong to.}

\item{grp_size}{A (1 x J) row vector indicating the number of covariates in each group.}

\item{grp_size_cs}{A (1 x J) row vector that is the cumulative sum of grp_size (indicating the indicies where each group ends).}

\item{alpha_inits}{A (K x 1) column vector containing initial values for the regression coefficients corresponding to C.}

\item{beta_inits}{A (M x 1) column vector containing initial values for the regression coefficients corresponding to X.}

\item{lambda_sq_inits}{A (M x 1) column vector containing initial values for the local shrinkage parameters.}

\item{gamma_sq_inits}{A (J x 1) column vector containing initial values for the group shrinkage parameters.}

\item{eta_inits}{A (J x 1) column vector containing initial values for the mixing parameters.}

\item{mu_init}{Initial value for the intercept parameter (double).}

\item{tau_sq_init}{Initial value for the global shrinkage parameter (double).}

\item{sigma_sq_init}{Initial value for the residual variance (double).}

\item{psi_sq_init}{Initial value for the variailbity in the alphas (double).}

\item{nu_init}{Initial value for the augmentation variable (double).}

\item{n_burn_in}{The number of burn-in samples (integer).}

\item{n_samples}{The number of posterior draws (integer).}

\item{n_thin}{The thinning interval (integer).}

\item{error_tol}{Parameter that controls numerical stability of the algorithm (double).}

\item{a1}{Shape parameter for the inverse gamma hyperprior on psi_sq (double).}

\item{b1}{Scale parameter for the inverse gamma hyperprior on psi_sq (double).}

\item{a2}{Shape parameter for the gamma hyperprior on the etas (double).}

\item{b2}{Scale parameter for the gamma hyperprior on the etas (double).}

\item{verbose}{Boolean value which indicates whether or not to print the progress of the Gibbs sampler.}
}
\value{
A list containing the posterior draws of (1) the intercept term (mus) (2) the regression coefficients (alphas and betas) (3) the individual shrinkage parameters (lambda_sqs) (4) the group shrinkage parameters (gamma_sqs) (5) the global shrinkage parameter (tau_sqs) (6) the residual error variance (sigma_sqs) (7) the mixing parameter (etas) and (8) the variance of the prior on alpha (psi_sqs). The list also contains the specified hyperparameters (hyperparam_a1, hyperparam_b1, hyperparam_a2, and hyperparam_b2), details regarding the dataset (X, C, Y, grp_idx), and Gibbs sampler details (n_burn_in, n_samples, and n_thin).
}
\description{
An Rcpp function that implements a Gibbs sampler for the grouped horseshoe model.
}
